migrate screens to features and migrate data repository to core-data module. finish the modularization.
