+++ app/build.gradle.kts
@@ -62,8 +62,11 @@ dependencies {
     implementation(project(":core:core-common"))
     implementation(project(":core:core-network"))
     implementation(project(":core:core-database"))
+    implementation(project(":core:core-data"))
     implementation(project(":core:core-uikit"))
     implementation(project(":domain"))
+    implementation(project(":features:feature-pokemon-list"))
+    implementation(project(":features:feature-pokemon-details"))
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
