--- app/src/main/java/dev/whysoezzy/pokemon/di/RepositoryModule.kt
@@ -1,12 +1,0 @@
-package dev.whysoezzy.pokemon.di
-
-import dev.whysoezzy.domain.repository.PokemonRepository
-import dev.whysoezzy.pokemon.data.repository.PokemonRepositoryImpl
-//import dev.whysoezzy.pokemon.domain.repository.PokemonRepository
-import org.koin.dsl.module
-
-val repositoryModule = module {
-    single<PokemonRepository> {
-        PokemonRepositoryImpl(get(), get())
-    }
-}
