+++ domain/build.gradle.kts
@@ -13,5 +13,5 @@ kotlin {
 }
 dependencies {
     implementation(libs.kotlinx.coroutines.core)
-
+    implementation(libs.koin.core)
 }
