+++ features/feature-pokemon-details/src/main/java/dev/whysoezzy/feature_pokemon_details/components/sections/PokemonPhysicalStatsSection.kt
@@ -1,0 +1,63 @@
+package dev.whysoezzy.feature_pokemon_details.components.sections
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Star
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import dev.whysoezzy.core_uikit.components.stats.PhysicalStatItem
+
+@Composable
+fun PokemonPhysicalStatsSection(
+    height: Int,
+    weight: Int,
+    modifier: Modifier = Modifier
+) {
+    Card(
+        modifier = modifier.fillMaxWidth(),
+        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
+    ) {
+        Column(
+            modifier = Modifier.padding(16.dp)
+        ) {
+            Text(
+                text = "Физические характеристики",
+                style = MaterialTheme.typography.titleLarge,
+                fontWeight = FontWeight.Bold
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceEvenly
+            ) {
+                PhysicalStatItem(
+                    icon = Icons.Default.Star,
+                    label = "Рост",
+                    value = "${height / 10.0} м",
+                    modifier = Modifier.weight(1f)
+                )
+
+                PhysicalStatItem(
+                    icon = Icons.Default.Star,
+                    label = "Вес",
+                    value = "${weight / 10.0} кг",
+                    modifier = Modifier.weight(1f)
+                )
+            }
+        }
+    }
+}
