+++ features/feature-pokemon-details/src/test/java/dev/whysoezzy/feature_pokemon_details/ExampleUnitTest.kt
@@ -1,0 +1,16 @@
+package dev.whysoezzy.feature_pokemon_details
+
+import org.junit.Assert.assertEquals
+import org.junit.Test
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
