+++ features/feature-pokemon-list/.gitignore
@@ -1,0 +1,1 @@
+/build
