+++ settings.gradle.kts
@@ -26,3 +26,7 @@ include("core:core-network")
 include(":core:core-common")
 include(":core:core-database")
 include(":core:core-uikit")
+include(":features:feature-pokemon-list")
+include(":features:feature-pokemon-details")
+//include(":core:core-data")
+include(":core:core-data")
